x64 engine and game build, frame census, and the software renderer's hooks
The x64 build (generated projects neo/doom-x64.sln, no editors): inline x87
and cpuid assembly replaced, pointers no longer squeezed into ints (heap,
handles, window procedures, GUI expression operands), demo and savegame
pointer flags through idFile::ReadPtrFlag / WritePtrFlag, the game's event
system and script VM made correct where a pointer is wider than a float,
GUI fonts loading again (the file size was compared with a struct that
holds a pointer), per-monitor DPI awareness. All of it compiles to the old
code on Win32.

Instrumentation: r_frameLog (tr_census.cpp, a per-frame census of the
command list), com_demoShotFrames and friends, timeDemoQuit .. twice,
r_showLightCount 4, crash.txt with a symbolized stack (win_crash.cpp), PDBs
in Release.

The software renderer's seam, all behind ID_SW_RENDERER (x64 only): the
back-end switch in R_IssueRenderCommands, CPU twins in the image loader, the
vertex cache kept in CPU memory, screenshots and captures reading the
software frame, r_swRenderer 2 without OpenGL (window without a context, the
generated null qgl table win_qgl_null.inl).

+++ neo/framework/Common.cpp
@@ -193,7 +193,7 @@ private:
 	idStrList					warningList;
 	idStrList					errorList;
 
-	int							gameDLL;
+	intptr_t					gameDLL;
 
 	idLangDict					languageDict;
